# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Start-Up India API",
    "description": "Start-Up India API Documentation"
  },
  "host": "localhost:443",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/startup/count/{from}/{to}": {
      "get": {
        "tags": [
          "Counts"
        ],
        "description": "Endpoint for date range wise count - India level",
        "parameters": [
          {
            "name": "from",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "to",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/startup/count/all": {
      "get": {
        "tags": [
          "Counts"
        ],
        "description": "Get India level startup count from inception",
        "parameters": [],
        "responses": {}
      }
    },
    "/startup/stages/{state}": {
      "get": {
        "tags": [
          "Business"
        ],
        "description": "Get state-wise stages",
        "parameters": [
          {
            "name": "state",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/startup/sectors/{state}": {
      "get": {
        "tags": [
          "Business"
        ],
        "description": "Get state-wise sectors",
        "parameters": [
          {
            "name": "state",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/startup/recognisedcount/all": {
      "get": {
        "tags": [
          "Counts"
        ],
        "description": "Count of recognised startups",
        "parameters": [],
        "responses": {}
      }
    },
    "/startup/dpiit/states": {
      "get": {
        "tags": [
          "Geography"
        ],
        "description": "List of all dpiit states",
        "parameters": [],
        "responses": {}
      }
    },
    "/startup/states": {
      "get": {
        "tags": [
          "Geography"
        ],
        "description": "List of all state with state id",
        "parameters": [],
        "responses": {}
      }
    },
    "/startup/districts/{stateId}": {
      "get": {
        "tags": [
          "Geography"
        ],
        "description": "List of all districts by state id",
        "parameters": [
          {
            "name": "stateId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/startup/industry/all": {
      "get": {
        "tags": [
          "Industry"
        ],
        "description": "List of all industries category in India",
        "parameters": [],
        "responses": {}
      }
    },
    "/startup/subIndustry/{industryId}": {
      "get": {
        "tags": [
          "Industry"
        ],
        "description": "List of sub-industries by industry id",
        "parameters": [
          {
            "name": "industryId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/data/table/{from}/{to}": {
      "get": {
        "tags": [
          "Data Tables"
        ],
        "description": "State-wise data table",
        "parameters": [
          {
            "name": "from",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "to",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {}
}